+++ R/check_heteroscedasticity.R
@@ -46,7 +46,6 @@ check_heteroskedasticity <- check_heteroscedasticity
 
 #' @export
 check_heteroscedasticity.default <- function(x, ...) {
-  # check for valid input
   .is_model_valid(x)
 
   # only for linear models