+++ R/check_predictions.R
@@ -222,7 +222,9 @@ check_predictions.BFBayesFactor <- function(object,
   if (isTRUE(is.na(re_formula))) {
     yy <- everything_we_need[["y_pred_marginal"]]
   } else {
-    if (!is.null(re_formula)) warning("re_formula can only be NULL or NA", call. = FALSE)
+    if (!is.null(re_formula)) {
+      insight::format_warning("`re_formula` can only be `NULL` or `NA`.")
+    }
     yy <- everything_we_need[["y_pred"]]
   }